Move implementation of ResourcePoolSourceData to separate file
