+++ Gems/Atom/RPI/Code/atom_rpi_edit_files.cmake
@@ -55,6 +55,7 @@ set(FILES
     Source/RPI.Edit/Material/MaterialFunctorSourceDataRegistration.cpp
     Source/RPI.Edit/Material/MaterialPipelineSourceData.cpp
     Source/RPI.Edit/Material/MaterialUtils.cpp
+    Source/RPI.Edit/ResourcePool/ResourcePoolSourceData.cpp
     Source/RPI.Edit/Shader/ShaderSourceData.cpp
     Source/RPI.Edit/Shader/ShaderVariantListSourceData.cpp
     Source/RPI.Edit/Shader/ShaderVariantAssetCreator.cpp
